+++ lib/src/x.dart
@@ -50,17 +50,19 @@ extension DioX on Dio {
 }
 
 extension FutureTioResponseX<R, E> on Future<TioResponse<R, E>> {
-  Future<T> map<T>({
-    required T Function(TioSuccess<R, E> success) success,
-    required T Function(TioFailure<R, E> failure) failure,
+  Future<K> map<K>({
+    required K Function(TioSuccess<R, E> success) success,
+    required K Function(TioFailure<R, E> failure) failure,
   }) =>
       then((response) => response.map(success: success, failure: failure));
 
-  Future<T> when<T>({
-    required T Function(R result) success,
-    required T Function(E error) failure,
+  Future<K> when<K>({
+    required K Function(R result) success,
+    required K Function(E error) failure,
   }) =>
       then((response) => response.when(success: success, failure: failure));
 
   Future<R> unwrap() => then((response) => response.requireResult);
+
+  Future<R?> maybeUnwrap() => then((response) => response.maybeResult);
 }